Hooks to use netket with single precision (#1544)
Remove hard-coded float64 constants and precision throughout netket and make it use the default jax value (float32/float64), so that it does not raise warnings if double precision is disabled.

Also, adds an env variable to disable double precision in netket.

+++ netket/sampler/base.py
@@ -68,7 +68,7 @@ class Sampler(abc.ABC):
     machine_pow: int = struct.field(default=2)
     """The power to which the machine should be exponentiated to generate the pdf."""
 
-    dtype: DType = struct.field(pytree_node=False, default=np.float64)
+    dtype: DType = struct.field(pytree_node=False, default=float)
     """The dtype of the states sampled."""
 
     def __pre_init__(
